+++ chocolatey-visualstudio.extension/extensions/Get-VSChannelUri.ps1
@@ -7,7 +7,7 @@ function Get-VSChannelUri
     )
 
     $manifestUri = $null
-    $success = $ChannelId -match '^VisualStudio\.(?<version>\d+)\.(?<kind>\w+)$' # VisualStudio.15.Release
+    $success = $ChannelId -match '^VisualStudio\.(?<version>\d+)\.(?<kind>[\w\.0-9]+)$' # VisualStudio.15.Release, VisualStudio.17.Release.LTSC.17.4
     if ($success)
     {
         $kind = switch ($Matches['kind'])
@@ -21,7 +21,9 @@ function Get-VSChannelUri
     }
     else
     {
-        Write-Error "Channel id '$ChannelId' does not match the expected pattern and cannot be used to compute the channel manifest URI"
+        $msg = "Channel id '$ChannelId' does not match the expected pattern and cannot be used to compute the channel manifest URI"
+        Write-Debug $msg
+        Write-Error $msg
     }
 
     return $manifestUri